project-service/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
alembic==1.10.3
python-jose==3.3.0
pika==1.3.1
httpx==0.23.3